@@ -12,49 +12,55 @@
     "build:webpack": "cross-env NODE_ENV=production encore production",
     "build": "run-s clean build:*"
   },
+  "browserslist": [
+    "> 0.5%",
+    "last 2 versions",
+    "Firefox ESR",
+    "not dead"
+  ],
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "devDependencies": {
-    "@11ty/eleventy": "^0.11.0",
-    "@11ty/eleventy-img": "^0.3.0",
-    "@babel/core": "^7.10.2",
-    "@babel/plugin-proposal-class-properties": "^7.10.1",
-    "@babel/plugin-proposal-object-rest-spread": "^7.10.1",
-    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/plugin-transform-runtime": "^7.10.1",
-    "@babel/preset-env": "^7.10.2",
-    "@symfony/webpack-encore": "^0.31.0",
-    "@tailwindcss/custom-forms": "^0.2.1",
-    "autoprefixer": "^9.8.6",
-    "babel-loader": "^8.1.0",
-    "cross-env": "^7.0.2",
-    "css-loader": "^4.3.0",
-    "file-loader": "^6.1.0",
-    "jsdom": "^16.4.0",
-    "lodash": "^4.17.20",
-    "netlify-plugin-image-optim": "^0.3.0",
-    "netlify-plugin-inline-critical-css": "^1.1.3",
-    "npm-run-all": "^4.1.5",
-    "postcss": "^8.1.1",
-    "postcss-loader": "^3.0.0",
-    "resolve-url-loader": "3.1.1",
-    "rimraf": "^3.0.2",
-    "tailwindcss": "^1.8.10",
-    "tailwindcss-aspect-ratio": "^3.0.0",
-    "tailwindcss-responsive-embed": "^1.0.0",
-    "url-loader": "^4.1.0",
-    "vue": "^2.6.12",
-    "vue-loader": "^15.9.3",
-    "vue-template-compiler": "^2.6.12",
-    "webpack-notifier": "^1.8.0"
-  },
   "dependencies": {
-    "@fortawesome/fontawesome-free": "^5.15.0",
+    "@11ty/eleventy": "^1.0.0-beta.10",
+    "@11ty/eleventy-img": "^1.0.0",
+    "@11ty/eleventy-upgrade-help": "^1.0.1",
+    "@babel/core": "^7.16.7",
+    "@babel/plugin-proposal-class-properties": "^7.16.7",
+    "@babel/plugin-proposal-object-rest-spread": "^7.16.7",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+    "@babel/plugin-transform-runtime": "^7.16.7",
+    "@babel/preset-env": "^7.16.7",
+    "@fortawesome/fontawesome-free": "^5.15.4",
+    "@symfony/webpack-encore": "^1.7.0",
+    "@tailwindcss/aspect-ratio": "^0.4.0",
+    "@tailwindcss/forms": "^0.4.0",
+    "@tailwindcss/line-clamp": "^0.3.0",
+    "@tailwindcss/typography": "^0.5.0",
+    "@vue/compiler-sfc": "^3.2.26",
+    "autoprefixer": "^10.4.1",
+    "babel-loader": "^8.2.3",
+    "cross-env": "^7.0.3",
+    "css-loader": "^6.5.1",
     "element-closest": "^3.0.2",
-    "glightbox": "^3.0.5",
-    "luxon": "^1.25.0",
-    "netlify-plugin-minify-html": "^0.3.0",
-    "swiper": "^6.3.2"
+    "file-loader": "^6.2.0",
+    "glightbox": "^3.1.0",
+    "jsdom": "^19.0.0",
+    "lodash": "^4.17.21",
+    "luxon": "^2.3.0",
+    "netlify-plugin-image-optim": "^0.4.0",
+    "netlify-plugin-inline-critical-css": "^2.0.0",
+    "netlify-plugin-minify-html": "^0.3.1",
+    "npm-run-all": "^4.1.5",
+    "postcss": "^8.4.5",
+    "postcss-loader": "^6.2.1",
+    "resolve-url-loader": "4.0.0",
+    "rimraf": "^3.0.2",
+    "swiper": "^7.4.1",
+    "tailwindcss": "^3.0.10",
+    "url-loader": "^4.1.1",
+    "vue": "^3.2.26",
+    "vue-loader": "^16.0",
+    "webpack-notifier": "^1.15.0"
   }
 }
