@@ -1,6 +1,6 @@
 
-
 import '../css/app.css';
+import './lazyload';
 import './menu';
 // import './contact';
 import './gallery';
