@@ -1,9 +1,12 @@
-import 'element-closest';
-
 document.addEventListener('DOMContentLoaded', (event) => {
     document.querySelector('.menu-toggle').addEventListener('click', event => {
         event.preventDefault();
-        document.querySelector('.main-menu').classList.toggle('hidden');
+        const menu = document.querySelector('.main-menu');
+        if (menu.classList.contains('hidden')) {
+            menu.classList.remove('hidden');
+        } else {
+            menu.classList.add('hidden');
+        }
     });
 
     document.addEventListener('click', event => {
