@@ -1,15 +1,18 @@
-import Swiper, { Navigation, Pagination, Autoplay } from 'swiper';
+import Swiper, { Navigation, Pagination, Autoplay, Lazy } from 'swiper';
 // import Swiper styles
-import 'swiper/swiper-bundle.css';
-
-Swiper.use([Navigation, Pagination, Autoplay]);
+import 'swiper/css';
+import 'swiper/css/navigation';
+import 'swiper/css/pagination';
+import 'swiper/css/autoplay';
+import 'swiper/css/lazy';
 
 (function () {
     document.addEventListener('DOMContentLoaded', function () {
         let sliders = document.querySelectorAll('[data-slider]');
         if(sliders.length > 0){
-            for(var i = 0; i < sliders.length; i++){
+            for(let i = 0; i < sliders.length; i++){
                 const swiper = new Swiper(sliders[i],{
+                    modules: [Navigation, Pagination, Autoplay, Lazy],
                     loop: true,
                     speed: 500,
                     effect: 'slide',
@@ -18,6 +21,10 @@ Swiper.use([Navigation, Pagination, Autoplay]);
                     autoplay: {
                         delay: 7500,
                         disableOnInteraction: true,
+                    },
+                    lazy: {
+                        loadPrevNext: true,
+                        loadOnTransitionStart: true,
                     },
                     pagination: {
                         el: '.swiper-pagination',
